MSR 4.0 – Organised and asynchronous IO, Stack health tracking
MobStackerReloaded now tracks the health of the entire stack, which is
affected by AOE damage. This way, players can deal damage to the entire
stack and kill it in whole, and the drops / XP will be multiplied on
death.

Furthermore, MSR now stores a flatfile for each mob, which reduces the
damage from possible data corruption. Additionally, these files are
loaded asynchronously on startup if they are numerous.